move quix runner to the platform folder
