+++ src/StreamingDataFrames/streamingdataframes/platforms/quix/__init__.py
@@ -1,2 +1,3 @@
 from .api import *
 from .config import *
+from .runner import QuixRunner
